Figma plugin update to export JSON data
Adds code to the plugin that gathers icon info in Figma, puts it in an object, stringifies it and shows it as a text field. Idea is that this icon index can be copied and submitted as a JSON file to Github, and then used in a landing page to render icons. This centralizes text info about icons in the Figma file, as the component name and description can be used.

Seems like an easy way to keep things organized.